backend/requirements.txt
flask
tensorflow
numpy
Pillow
matplotlib
